1. 插件集在service退出时, 需要先卸载动态库, 如果有协程未退出，要等协程全部完成后退出 2. fix 编译失败

+++ OptionComponent/OptionComp/storage/MongoDB/Impl/MongoDbMgr.cpp
@@ -578,7 +578,7 @@ KERNEL_NS::CoTask<bool> MongoDbMgr::AddData(KERNEL_NS::LibString dbName, KERNEL_
                 auto &firstStr = kv.first;
                 if(kv.second.IsBriefData())
                 {
-                    fullDoc.append(bsoncxx::builder::basic::kvp(firstStr.GetRaw(), kv.second.AsInt64()));
+                    fullDoc.append(bsoncxx::builder::basic::kvp(firstStr.GetRaw(), static_cast<std::int64_t>(kv.second.AsInt64())));
                 }
                 else if(kv.second.IsStr())
                 {
@@ -701,7 +701,7 @@ KERNEL_NS::CoTask<bool> MongoDbMgr::DelData(KERNEL_NS::LibString dbName, KERNEL_
                 auto &key = kv.first;
                 if(kv.second.IsBriefData())
                 {
-                    fullKv.append(bsoncxx::builder::basic::kvp(key.GetRaw(), kv.second.AsInt64()));
+                    fullKv.append(bsoncxx::builder::basic::kvp(key.GetRaw(),  static_cast<std::int64_t>(kv.second.AsInt64())));
                 }
                 else if(kv.second.IsStr())
                 {
